models/mal-file/requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
yara-python
pefile